Online store built with Flask/requirements.txt
bidict==0.22.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==3.2.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyMySQL==1.0.2
PyMySQL==1.0.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2022.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0